requirements.txt
flask
tensorflow
numpy
pillow
opencv-python
scipy
streamlit
